Revamp all droppedCard methods so that they take cards off of the source stack
* Removed friend classes
* Simplified the code in the drop handlers, by moving the takeCard code to where it belongs, in the stacks.
* created new takeCard(Card* ) methods in all stacks
* This is still kinda ugly, should move things off the stack during the drag instead of doing it in the drop. but it's working well

* This also  fixed problems with the hand messing up the sorted stacks when it recovers the waste pile

+++ QtSolitaire/deck.cpp
@@ -39,7 +39,7 @@ void Deck::addCard(Card* card)
 {
     RandomStack::addCard(card);
     double x = (double)((mCards.size()-1) % 13) * 30.0;
-    int y = (double)((mCards.size()-1) / 13) * 30.0;
+    double y = (double)((mCards.size()-1) / 13) * 30.0;
     card->setPos(x, y);
 }
